// src/types/fileRow.ts
export interface FileRow {
  url?: string;
  isValid?: boolean;
  status?: number;
  isXml?: boolean;
  typeValue?: string;
  styleContent?: string;
  과목코드?: string;
  학년?: string | number;
  학기?: string | number;
  단원순서?: string | number;
  학년E?: string | number;
  목차일련번호?: string | number;
}